requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
httpie==3.2.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
Pygments==2.18.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
